[Feat] 프로필 tagList에서 cell 클릭 시 Snap으로 화면전환 구현
